+++ src/Symfony/Bridge/Doctrine/Tests/Logger/DbalLoggerTest.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Bridge\Doctrine\Tests\Logger;
 
-
 class DbalLoggerTest extends \PHPUnit_Framework_TestCase
 {
     /**