+++ src/Symfony/Component/ClassLoader/UniversalClassLoader.php
@@ -242,7 +242,7 @@ class UniversalClassLoader
      * Loads the given class or interface.
      *
      * @param string $class The name of the class
-     * 
+     *
      * @return Boolean|null True, if loaded
      */
     public function loadClass($class)