+++ src/Symfony/Component/Config/Loader/Loader.php
@@ -47,10 +47,12 @@ abstract class Loader implements LoaderInterface
      *
      * @param mixed  $resource A Resource
      * @param string $type     The resource type
+     *
+     * @return mixed
      */
     public function import($resource, $type = null)
     {
-        $this->resolve($resource)->load($resource, $type);
+        return $this->resolve($resource)->load($resource, $type);
     }
 
     /**