+++ src/Symfony/Component/CssSelector/Node/FunctionNode.php
@@ -209,7 +209,6 @@ class FunctionNode implements NodeInterface
         $xpath->addCondition(sprintf('contains(string(.), %s)', XPathExpr::xpathLiteral($expr)));
 
         // FIXME: Currently case insensitive matching doesn't seem to be happening
-
         return $xpath;
     }
 
@@ -264,7 +263,6 @@ class FunctionNode implements NodeInterface
 
         if (false === strpos($s, 'n')) {
             // Just a b
-
             return array(0, intval((string) $s));
         }