+++ src/Symfony/Component/Form/Extension/Core/DataTransformer/DateTimeToLocalizedStringTransformer.php
@@ -32,12 +32,12 @@ class DateTimeToLocalizedStringTransformer extends BaseDateTimeTransformer
      *
      * @see BaseDateTimeTransformer::formats for available format options
      *
-     * @param string             $inputTimezone  The name of the input timezone
-     * @param string             $outputTimezone The name of the output timezone
-     * @param integer            $dateFormat     The date format
-     * @param integer            $timeFormat     The time format
-     * @param \IntlDateFormatter $calendar       An \IntlDateFormatter instance
-     * @param string             $pattern        A pattern to pass to \IntlDateFormatter
+     * @param string  $inputTimezone  The name of the input timezone
+     * @param string  $outputTimezone The name of the output timezone
+     * @param integer $dateFormat     The date format
+     * @param integer $timeFormat     The time format
+     * @param integer $calendar       One of the \IntlDateFormatter calendar constants
+     * @param string  $pattern        A pattern to pass to \IntlDateFormatter
      *
      * @throws UnexpectedTypeException If a format is not supported
      * @throws UnexpectedTypeException if a timezone is not a string