+++ src/Symfony/Component/Form/Extension/Core/Type/DateType.php
@@ -180,6 +180,10 @@ class DateType extends AbstractType
             );
         };
 
+        $format = function (Options $options) {
+            return $options['widget'] === 'single_text' ? DateType::HTML5_FORMAT : DateType::DEFAULT_FORMAT;
+        };
+
         // BC until Symfony 2.3
         $modelTimezone = function (Options $options) {
             return $options['data_timezone'];
@@ -196,7 +200,7 @@ class DateType extends AbstractType
             'days'           => range(1, 31),
             'widget'         => 'choice',
             'input'          => 'datetime',
-            'format'         => self::HTML5_FORMAT,
+            'format'         => $format,
             'model_timezone' => $modelTimezone,
             'view_timezone'  => $viewTimezone,
             // Deprecated timezone options